+++ core-ng/src/main/java/core/framework/internal/db/InsertQuery.java
@@ -4,30 +4,21 @@ package core.framework.internal.db;
  * @author neo
  */
 final class InsertQuery<T> {
-    final String sql;
-    final String upsertClause;
+    final String insertSQL;
+    final String insertIgnoreSQL;
+    final String upsertSQL;
     final String generatedColumn;
     private final InsertQueryParamBuilder<T> paramBuilder;
 
-    InsertQuery(String sql, String upsertClause, String generatedColumn, InsertQueryParamBuilder<T> paramBuilder) {
-        this.sql = sql;
-        this.upsertClause = upsertClause;
+    InsertQuery(String insertSQL, String insertIgnoreSQL, String upsertSQL, String generatedColumn, InsertQueryParamBuilder<T> paramBuilder) {
+        this.insertSQL = insertSQL;
+        this.insertIgnoreSQL = insertIgnoreSQL;
+        this.upsertSQL = upsertSQL;
         this.generatedColumn = generatedColumn;
         this.paramBuilder = paramBuilder;
     }
 
     Object[] params(T entity) {
         return paramBuilder.params(entity);
-    }
-
-    String insertIgnoreSQL() {
-        // due to insert ignore will be used less frequently, so not to build sql in advance to reduce memory footprint
-        // replace first INSERT with INSERT IGNORE
-        // new StringBuilder(str) will reserve str.length+16 as capacity, so insert will not trigger expansion
-        return new StringBuilder(sql).insert(6, " IGNORE").toString();
-    }
-
-    String upsertSQL() {
-        return sql + upsertClause;
     }
 }