+++ core-ng/src/main/java/core/framework/internal/db/cloud/CloudAuthProvider.java
@@ -4,8 +4,6 @@ package core.framework.internal.db.cloud;
  * @author neo
  */
 public interface CloudAuthProvider {
-
-
     String user();
 
     String accessToken();