+++ core-ng/src/test/java/core/framework/internal/db/SelectQueryTest.java
@@ -1,6 +1,5 @@
 package core.framework.internal.db;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -11,27 +10,30 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author neo
  */
 class SelectQueryTest {
-    private SelectQuery<AssignedIdEntity> selectQuery;
-
-    @BeforeEach
-    void createSelectQuery() {
-        selectQuery = new SelectQuery<>(AssignedIdEntity.class);
-    }
-
     @Test
     void getSQL() {
-        assertThat(selectQuery.getSQL).isEqualTo("SELECT id, string_field, int_field, big_decimal_field, date_field, zoned_date_time_field FROM assigned_id_entity WHERE id = ?");
+        SelectQuery<AssignedIdEntity> query = new SelectQuery<>(AssignedIdEntity.class, Dialect.MYSQL);
+        assertThat(query.getSQL).isEqualTo("SELECT id, string_field, int_field, big_decimal_field, date_field, zoned_date_time_field FROM assigned_id_entity WHERE id = ?");
     }
 
     @Test
     void fetchSQL() {
-        String sql = selectQuery.fetchSQL(new StringBuilder("string_field = ?"), "int_field ASC", 4, 10);
+        SelectQuery<AssignedIdEntity> query = new SelectQuery<>(AssignedIdEntity.class, Dialect.MYSQL);
+        String sql = query.fetchSQL(new StringBuilder("string_field = ?"), "int_field ASC", 4, 10);
         assertThat(sql).isEqualTo("SELECT id, string_field, int_field, big_decimal_field, date_field, zoned_date_time_field FROM assigned_id_entity WHERE string_field = ? ORDER BY int_field ASC LIMIT ?,?");
     }
 
     @Test
+    void fetchSQLWithPostgreSQL() {
+        SelectQuery<AssignedIdEntity> query = new SelectQuery<>(AssignedIdEntity.class, Dialect.POSTGRESQL);
+        String sql = query.fetchSQL(new StringBuilder("string_field = ?"), "int_field ASC", 4, 10);
+        assertThat(sql).isEqualTo("SELECT id, string_field, int_field, big_decimal_field, date_field, zoned_date_time_field FROM assigned_id_entity WHERE string_field = ? ORDER BY int_field ASC OFFSET ? LIMIT ?");
+    }
+
+    @Test
     void params() {
-        Object[] params = selectQuery.params(List.of("value"), null, 100);
+        SelectQuery<AssignedIdEntity> query = new SelectQuery<>(AssignedIdEntity.class, Dialect.MYSQL);
+        Object[] params = query.params(List.of("value"), null, 100);
 
         assertThat(params).containsExactly("value", 0, 100);    // default skip should be 0
     }